# design/AwsProj/internal/app/handler/element.go
package handler

import (
	"net/http"
	"strconv"
	"strings"

	"AwsProj/internal/app/ds"
	"AwsProj/internal/app/service"

	"github.com/gin-gonic/gin"
	"github.com/sirupsen/logrus"
)

// GetAllElements получение всех элементов

// GetAllElements godoc
// @Summary Get all elements
// @Description Get list of all elements with optional search by name
// @Tags elements
// @Accept json
// @Produce json
// @Param query query string false "Search query by element name"
// @Success 200 {object} map[string]interface{} "Success response with elements list"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /elements [get]
func (h *Handler) GetAllElements(ctx *gin.Context) {
	var elements []ds.Elements
	var err error

	search := ctx.Query("query")
	if search == "" {
		elements, err = h.Repository.GetAllElements()
	} else {
		elements, err = h.Repository.SearchElementByName(search)
	}

	if err != nil {
		ctx.JSON(http.StatusInternalServerError, ErrorResponse{
			Success: false,
			Error:   "Failed to get elements",
			Message: err.Error(),
		})
		logrus.Error(err)
		return
	}

	// Преобразуем в JSON-ответ
	var response []ElementResponse
	for _, elem := range elements {
		response = append(response, ElementResponse{
			ID:            elem.ID,
			Name:          elem.Name,
			Description:   elem.Description,
			Ph:            elem.Ph,
			Concentration: elem.Concentration,
			Image:         elem.Img,
		})
	}

	ctx.JSON(http.StatusOK, SuccessResponse{
		Success: true,
		Data: gin.H{
			"items": response,
			"total": len(response),
			"query": search,
		},
	})
}

// GetElementById получение элемента по ID

// GetElementById godoc
// @Summary Get element by ID
// @Description Retrieve a specific element by its ID
// @Tags elements
// @Accept json
// @Produce json
// @Param id path int true "Element ID"
// @Success 200 {object} SuccessResponse "Element found"
// @Failure 400 {object} ErrorResponse "Invalid ID format"
// @Failure 404 {object} ErrorResponse "Element not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /elements/{id} [get]
func (h *Handler) GetElementById(ctx *gin.Context) {
	strId := ctx.Param("id")
	id, err := strconv.Atoi(strId)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, ErrorResponse{
			Success: false,
			Error:   "Invalid ID format",
			Message: "ID must be a number",
		})
		logrus.Error("Invalid ID format:", err)
		return
	}

	element, err := h.Repository.GetElementByID(id)
	if err != nil {
		ctx.JSON(http.StatusNotFound, ErrorResponse{
			Success: false,
			Error:   "Element not found",
			Message: err.Error(),
		})
		logrus.Error("Element not found:", err)
		return
	}

	response := ElementResponse{
		ID:            element.ID,
		Name:          element.Name,
		Description:   element.Description,
		Ph:            element.Ph,
		Concentration: element.Concentration,
		Image:         element.Img,
	}

	ctx.JSON(http.StatusOK, SuccessResponse{
		Success: true,
		Data:    response,
	})
}

// CreateElement создание элемента (только для админов)

// CreateElement godoc
// @Summary Create new element
// @Description Create a new chemical element (Admin only)
// @Tags elements
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param request body CreateElementRequest true "Element data"
// @Success 201 {object} SuccessResponse "Element created successfully"
// @Failure 400 {object} ErrorResponse "Invalid request data or validation error"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 403 {object} ErrorResponse "Forbidden - Admin access required"
// @Failure 409 {object} ErrorResponse "Element already exists"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /elements [post]
func (h *Handler) CreateElement(ctx *gin.Context) {
	var req CreateElementRequest

	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, ErrorResponse{
			Success: false,
			Error:   "Invalid request data",
			Message: err.Error(),
		})
		return
	}

	// Конвертируем в service request
	serviceReq := &service.CreateElementRequest{
		Name:          req.Name,
		Description:   req.Description,
		Ph:            req.Ph,
		Concentration: req.Concentration,
	}

	result, err := h.ElementService.CreateElement(serviceReq)
	if err != nil {
		statusCode := http.StatusInternalServerError
		if strings.Contains(err.Error(), "уже существует") ||
			strings.Contains(err.Error(), "обязательно") ||
			strings.Contains(err.Error(), "диапазон") {
			statusCode = http.StatusBadRequest
		}

		ctx.JSON(statusCode, ErrorResponse{
			Success: false,
			Error:   "Failed to create element",
			Message: err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusCreated, SuccessResponse{
		Success: true,
		Message: result.Message,
		Data:    result,
	})
}

// UpdateElement обновление элемента (только для админов)

// UpdateElement godoc
// @Summary Update element
// @Description Update an existing chemical element (Admin only)
// @Tags elements
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Element ID"
// @Param request body UpdateElementRequest true "Updated element data"
// @Success 200 {object} SuccessResponse "Element updated successfully"
// @Failure 400 {object} ErrorResponse "Invalid request data or validation error"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 403 {object} ErrorResponse "Forbidden - Admin access required"
// @Failure 404 {object} ErrorResponse "Element not found"
// @Failure 409 {object} ErrorResponse "Element with this name already exists"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /elements/{id} [put]
func (h *Handler) UpdateElement(ctx *gin.Context) {
	strID := ctx.Param("id")
	id, err := strconv.ParseUint(strID, 10, 32)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, ErrorResponse{
			Success: false,
			Error:   "Invalid element ID",
			Message: "ID must be a number",
		})
		return
	}

	var req UpdateElementRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, ErrorResponse{
			Success: false,
			Error:   "Invalid request data",
			Message: err.Error(),
		})
		return
	}

	// Конвертируем в service request
	serviceReq := &service.UpdateElementRequest{
		Name:          req.Name,
		Description:   req.Description,
		Ph:            req.Ph,
		Concentration: req.Concentration,
	}

	result, err := h.ElementService.UpdateElement(int(id), serviceReq)
	if err != nil {
		statusCode := http.StatusInternalServerError
		if strings.Contains(err.Error(), "не найден") {
			statusCode = http.StatusNotFound
		} else if strings.Contains(err.Error(), "уже существует") ||
			strings.Contains(err.Error(), "диапазон") ||
			strings.Contains(err.Error(), "отрицательной") ||
			strings.Contains(err.Error(), "нет данных") {
			statusCode = http.StatusBadRequest
		}

		ctx.JSON(statusCode, ErrorResponse{
			Success: false,
			Error:   "Failed to update element",
			Message: err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, SuccessResponse{
		Success: true,
		Message: result.Message,
		Data:    result,
	})
}

// DeleteElement удаление элемента (только для админов)

// DeleteElement godoc
// @Summary Delete element
// @Description Delete a chemical element (Admin only)
// @Tags elements
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param id path int true "Element ID"
// @Success 200 {object} SuccessResponse "Element deleted successfully"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 403 {object} ErrorResponse "Forbidden - Admin access required"
// @Failure 404 {object} ErrorResponse "Element not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /elements/{id} [delete]
func (h *Handler) DeleteElement(ctx *gin.Context) {
	strID := ctx.Param("id")
	id, err := strconv.Atoi(strID)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, ErrorResponse{
			Success: false,
			Error:   "Invalid element ID",
			Message: "ID must be a number",
		})
		return
	}

	result, err := h.ElementService.DeleteElement(id)
	if err != nil {
		statusCode := http.StatusInternalServerError
		if strings.Contains(err.Error(), "не найден") {
			statusCode = http.StatusNotFound
		}

		ctx.JSON(statusCode, ErrorResponse{
			Success: false,
			Error:   "Failed to delete element",
			Message: err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, SuccessResponse{
		Success: true,
		Message: result.Message,
		Data:    result,
	})
}

// UploadImage загрузка изображения для элемента (только для админов)

// UploadImage godoc
// @Summary Upload element image
// @Description Upload an image for a chemical element (Admin only)
// @Tags elements
// @Accept multipart/form-data
// @Produce json
// @Security BearerAuth
// @Param id path int true "Element ID"
// @Param image formData file true "Image file (JPEG, PNG, GIF, WebP). Max 5MB"
// @Success 200 {object} SuccessResponse "Image uploaded successfully"
// @Failure 400 {object} ErrorResponse "Invalid file or element ID"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 403 {object} ErrorResponse "Forbidden - Admin access required"
// @Failure 404 {object} ErrorResponse "Element not found"
// @Failure 413 {object} ErrorResponse "File too large"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /elements/{id}/image [post]
func (h *Handler) UploadImage(ctx *gin.Context) {
	strID := ctx.Param("id")
	elementID, err := strconv.Atoi(strID)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, ErrorResponse{
			Success: false,
			Error:   "Invalid element ID",
			Message: "ID must be a number",
		})
		return
	}

	fileHeader, err := ctx.FormFile("image")
	if err != nil {
		ctx.JSON(http.StatusBadRequest, ErrorResponse{
			Success: false,
			Error:   "File not found",
			Message: "Image file is required",
		})
		return
	}

	result, err := h.ElementService.UploadImage(elementID, fileHeader)
	if err != nil {
		statusCode := http.StatusInternalServerError
		if strings.Contains(err.Error(), "не найден") {
			statusCode = http.StatusNotFound
		} else if strings.Contains(err.Error(), "размер") ||
			strings.Contains(err.Error(), "поддерживаются") {
			statusCode = http.StatusBadRequest
		}

		ctx.JSON(statusCode, ErrorResponse{
			Success: false,
			Error:   "Failed to upload image",
			Message: err.Error(),
		})
		return
	}

	ctx.JSON(http.StatusOK, SuccessResponse{
		Success: true,
		Message: result.Message,
		Data:    result,
	})
}